Project/SV/navigator-master/acni3shared/src/main/java/com/ni3/ag/adminconsole/domain/Map.java
/** * Copyright (c) 2009-2015 Social Vision GmbH. All rights reserved. */
package com.ni3.ag.adminconsole.domain;

import java.io.Serializable;

public class Map implements Serializable{
	private static final long serialVersionUID = 1L;
	private Integer id;
	private String name;

	public Integer getId(){
		return id;
	}

	public void setId(Integer id){
		this.id = id;
	}

	public String getName(){
		return name;
	}

	public void setName(String name){
		this.name = name;
	}

	@Override
	public boolean equals(java.lang.Object o){
		if (o == null)
			return false;
		if (!(o instanceof Map))
			return false;
		if (o == this)
			return true;
		Map dt = (Map) o;
		if (getId() == null || dt.getId() == null)
			return false;
		return getId().equals(dt.getId());
	}

}
